requirements_render.txt
# Minimal requirements for Render deployment
eel==0.16.0
google-generativeai
python-dotenv==1.0.0
requests==2.31.0